Enables cluster setup
This generates a cluster section in a rabbitmq.conf that
adds all the members of a current play into it, resulting
in nodes being added to a cluster automatically during bootsrap.
Adds following variables:
  rabbitmq_cluster - boolean, set to True to enable the cluster
  rabbitmq_erlang_cookie - string, unique Erlang cluster identifier

The proper generation of cluster section config depends on
'ansible_play_hosts' magic variable, that is found in ansible 2.2 or later.